==> _previous/clustering/sjvin/cluster_baselineCan2025/02-6_preparation_donnees_ses.ipynb <==
# Préparation des variables de ses

library(tidyverse)
library(factoextra)

# Charger les données
DataPilot <- readRDS("../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/01_pilot_2025.rds")

# Charger le fichier des variables
source("01-1_variables.R")

df_ses <- DataPilot %>%
  select(all_of(variables_ses))

# Boucle pour afficher la table pour chaque variable
for (v in variables_ses) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_ses[[v]]))
    cat("\n\n")
  }
}

### Transformations

Les variables age, éducation et income devraient être des variables ordinales

# Transformation de âge en variable ordinale

df_ses <- df_ses %>%
  mutate(
    ses_age = case_when(
      ses_ageGroup5Years %in% c("18_19", "20_24", "25_29", "30_34") ~ 0,
      ses_ageGroup5Years %in% c("35_39", "40_44", "45_49", "50_54") ~ 0.5,
      ses_ageGroup5Years %in% c("55_59", "60_64", "65_69", "70_74",
                                "75_79", "80_84", "85_89", "90_94",
                                "95_99", "100+") ~ 1,
      TRUE ~ NA_real_
    ) 
  ) %>%
  select(
    -ses_ageGroup5Years,
    -ses_age_4Cat
  )


# Transformation de educ en variable ordinale
#--------------------------------------------

df_ses$ses_educBHS <- NA
df_ses$ses_educBHS[df_ses$ses_educ_3Cat == 'educBHS'] <- 1
df_ses$ses_educBHS[df_ses$ses_educ_3Cat != 'educBHS'] <- 0
table(df_ses$ses_educBHS)

df_ses$ses_educPostHS <- NA
df_ses$ses_educPostHS[df_ses$ses_educ_3Cat == 'educPostHS'] <- 1
df_ses$ses_educPostHS[df_ses$ses_educ_3Cat != 'educPostHS'] <- 0
table(df_ses$ses_educPostHS)

df_ses$ses_educUniv <- NA
df_ses$ses_educUniv[df_ses$ses_educ_3Cat == 'educUniv'] <- 1
df_ses$ses_educUniv[df_ses$ses_educ_3Cat != 'educUniv'] <- 0
table(df_ses$ses_educUniv)

table(df_ses$ses_educ)

# Transformation de income en variable ordinale
#----------------------------------------------

df_ses$ses_incomeLow <- NA
df_ses$ses_incomeLow[df_ses$ses_income3Cat == 'Low'] <- 1
df_ses$ses_incomeLow[df_ses$ses_income3Cat != 'Low'] <- 0
table(df_ses$ses_incomeLow)

df_ses$ses_incomeMid <- NA
df_ses$ses_incomeMid[df_ses$ses_income3Cat == 'Mid'] <- 1
df_ses$ses_incomeMid[df_ses$ses_income3Cat != 'Mid'] <- 0
table(df_ses$ses_incomeMid)

df_ses$ses_incomeHigh <- NA
df_ses$ses_incomeHigh[df_ses$ses_income3Cat == 'High'] <- 1
df_ses$ses_incomeHigh[df_ses$ses_income3Cat != 'High'] <- 0
table(df_ses$ses_incomeHigh)


# Binariser ses_language
#-----------------------------------------

df_ses$ses_languageEnglish <- NA
df_ses$ses_languageEnglish[df_ses$ses_language == 'english'] <- 1
df_ses$ses_languageEnglish[df_ses$ses_language != 'english'] <- 0
table(df_ses$ses_languageEnglish)

df_ses$ses_languageFrench <- NA
df_ses$ses_languageFrench[df_ses$ses_language == 'french'] <- 1
df_ses$ses_languageFrench[df_ses$ses_language != 'french'] <- 0
table(df_ses$ses_languageFrench)

df_ses$ses_languageOther <- NA
df_ses$ses_languageOther[df_ses$ses_language == 'other'] <- 1
df_ses$ses_languageOther[df_ses$ses_language != 'other'] <- 0
table(df_ses$ses_languageOther)

# Binariser ses_ethnicityWB
#-----------------------------------------

df_ses$ses_ethnicityWhite <- NA
df_ses$ses_ethnicityWhite[df_ses$ses_ethnicityWB == 'white'] <- 1
df_ses$ses_ethnicityWhite[df_ses$ses_ethnicityWB != 'white'] <- 0
table(df_ses$ses_ethnicityWhite)

df_ses$ses_ethnicityMinority <- NA
df_ses$ses_ethnicityMinority[df_ses$ses_ethnicityWB != 'white'] <- 1
df_ses$ses_ethnicityMinority[df_ses$ses_ethnicityWB == 'white'] <- 0
table(df_ses$ses_ethnicityMinority)


# Binariser ses_sexOrientation
#-----------------------------------------

df_ses$ses_sexOrientationHetero <- NA
df_ses$ses_sexOrientationHetero[df_ses$ses_sexOrientation == 'heterosexual'] <- 1
df_ses$ses_sexOrientationHetero[df_ses$ses_sexOrientation != 'heterosexual'] <- 0
table(df_ses$ses_sexOrientationHetero)

df_ses$ses_sexOrientationQueer <- NA
df_ses$ses_sexOrientationQueer[df_ses$ses_sexOrientation %in% c('gay', 'bisexual', 'other')] <- 1
df_ses$ses_sexOrientationQueer[df_ses$ses_sexOrientation == 'heterosexual'] <- 0
table(df_ses$ses_sexOrientationQueer)

# Binariser ses_region
#-----------------------------------------

df_ses$ses_regionPrairies <- NA
df_ses$ses_regionPrairies[df_ses$ses_region == "prairie"] <- 1
df_ses$ses_regionPrairies[df_ses$ses_region != "prairie"] <- 0

df_ses$ses_regionBC <- NA
df_ses$ses_regionBC[df_ses$ses_region == "british_columbia"] <- 1
df_ses$ses_regionBC[df_ses$ses_region != "british_columbia"] <- 0

df_ses$ses_regionAtlantic <- NA
df_ses$ses_regionAtlantic[df_ses$ses_region == "atlantic"] <- 1
df_ses$ses_regionAtlantic[df_ses$ses_region != "atlantic"] <- 0

df_ses$ses_regionOntario <- NA
df_ses$ses_regionOntario[df_ses$ses_region == "ontario"] <- 1
df_ses$ses_regionOntario[df_ses$ses_region != "ontario"] <- 0

df_ses$ses_regionQuebec <- NA
df_ses$ses_regionQuebec[df_ses$ses_region == "quebec"] <- 1
df_ses$ses_regionQuebec[df_ses$ses_region != "quebec"] <- 0

df_ses$ses_regionTerritories <- NA
df_ses$ses_regionTerritories[df_ses$ses_region == "territories"] <- 1
df_ses$ses_regionTerritories[df_ses$ses_region != "territories"] <- 0

# Regrouper ses_region
table(df_ses$ses_region)
df_ses$ses_regionWest <- NA
df_ses$ses_regionWest[df_ses$ses_region %in% c("prairie", "british_columbia")] <- 1
df_ses$ses_regionWest[df_ses$ses_region %in% c("atlantic", "ontario", "quebec", "territories")] <- 0
table(df_ses$ses_regionWest)

variables_ses_clust <- c(
  "id",
  "ses_genderMale",
  "ses_age",
  "ses_languageEnglish",
  "ses_languageFrench",
  "ses_languageOther",
  "ses_educBHS",
  "ses_educPostHS",
  "ses_educUniv",
  "ses_incomeLow",
  "ses_incomeMid",
  "ses_incomeHigh",
  "ses_immigrant",
  "ses_ethnicityWhite",
  "ses_ethnicityMinority",
  "ses_sexOrientationHetero",
  "ses_sexOrientationQueer",
  "ses_regionPrairies",
  "ses_regionBC",
  "ses_regionAtlantic",
  "ses_regionOntario",
  "ses_regionQuebec",
  "ses_regionTerritories",
  "ses_regionWest"
)

# Boucle pour afficher la table pour chaque variable
for (v in variables_ses_clust) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_ses[[v]]))
    cat("\n\n")
  }
}

# Sauvegarder les données préparées
saveRDS(df_ses, file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/02_pilot_2025_ses.rds")

---

# SUGGESTION : Activité de Clustering

Si souhaité, effectuer un exercice de clustering avec ce sous-groupe de variables pour mieux saisir les données et leur relation. 

Cet exercice est purement exploratoire et sert à voir si les données peuvent être utilisées pour faire du clustering.
